requirements.txt
streamlit
google-cloud-firestore
vertexai
markdown2
xhtml2pdf